requirements.txt
asgiref==3.9.2
boto3==1.40.44
botocore==1.40.44
Django==5.2.7
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
jmespath==1.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
psycopg2-binary
python-dotenv
